Create an events plugin
Add a stream command which behaves much like the core stream-events
command except that it can additionally take a --ids switch which add
event ids to each event, and a --resume-after switch to start streaming
events at an older event.  This plugin captures events in memory, so the
events are lost on server restart, and the ids will reset.

Change-Id: I9343d04dd62eed582144a56f0e07342b7972e314