Complete Multi-master safety for the event plugin.
Use a new EventSequence to synchronize the event submissions with the
Head sequence.  This completes the last remaining piece to ensure that
the event plugin is multi-master safe.  The fsstore under
site_dir>/data/plugin/events/fsstore-2 must reside  on a cluster (likely
NFS) filesystem in order to share the FsStore among multiple nodes.

Change-Id: I92aa11811f28d5339d56484f5cdd9335795be749