Update patch set 1
Patch Set 1: Verified-1

I can't seem to generate the "merge-failed" event as of 2.12, in either core or the plugin.  Any thoughts on why?  I know the merge queue was eliminated, so the 'review --submit' command now returns with an error.  But it puts a merge failure comment on the change, so I would have expected it to create the event also.  The event still exists in core, and still appears to be called by MergeOp, so it is odd that I don't see it triggered anymore.

Patch-set: 1
Reviewer: Martin Fick <1010013@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=-1