Poll the filesystem for new events
In a multi-master environment events may be created and written
to the filestore from any master.  Without a notification system,
events written by other servers might only be seen by the current
server when the current server creates a new event.  If the traffic
on the current server is low, this delay may be significant and
unacceptable.

In order to get events from other masters output in a timely manner,
poll the filesystem to recongnize when new events may have been
written to the filestore and output them.  This makes events timely
in a multi-master environment without any notification system, and
makes it more robustly timely in systems with notifications.

Change-Id: Ic474f3ca639cc839b555f1ee82c97cc3e5882a89