Make event serialization encoding more explicit
Create a Fs.writeUtf8() method that explicitly writes using UTF8 instead
of relying on the default to be UTF8. Also rename the Fs.readFile()
method to readUtf8() to make it more obvious that it is deserializing
using UTF8 and to make it an obvious counterpart to the new writeUtf8()
method.

Change-Id: Id270a3f1b8b9dfc86d63f21ad1854b00476f3f87