Avoid read and updating the event sequences simultaneously
Add a java synchronized around the reads and updates to the sequence
values to ensure that a single process on the same host using the object
correctly does not alone cause NFS_silly_renames which can result in the
loss of the value file. Adjust the FsStoreTest to use a common FsStore
so that the synchronized will work across the different test threads.
Testing shows that with this change, the value file is never lost from
the code itself. In theory it can still be lost be reading the value
file externally or running this more than once on the same host.

Change-Id: Ifc7651f8123550d41faaeade16885244bf5101ac