Make reading event sequence files safer
Capture ArrayIndexOutOfBoundsException which is a RuntimeException and
therefore was missed. This exception can be thrown when reading a file
which is concurrently updated. This exception has been seen when the
value file was replaced (a normal occurrence for any updates).
Previously if this exception occurred, then the Gerrit operation causing
the event could fail and the event could be lost; this should no longer
be the case.

Change-Id: Id53f9529ba921d6d7ba6a32be1ece73987f91ec0