Support to run shell script tests in docker environment
Automate shell script tests to run in docker environment so that we can
reduce the manual effort for running these tests. In this setup there
are two containers, one for running gerrit server and another for
running tests. The gerrit server is hosted on gerrit-01 container and
the tests are run on run_tests container, default user which is used to
run docker tests is created by impersonating the 'admin' account which
is created when gerrit is initialized.

Change-Id: I0b4fdebdf9e27cf5250e45845b562bde58248d0e