test/docker: Use gerritcodereview docker image
Use a 2.14.21-ubuntu16 gerritcodereview image tag as the plugin
currently is compatible with 2.14 gerrit version.

Retain the options for specifying a gerrit.war and/or events.jar
but convert them to use local paths since the use cases for those
now are to specify locally-built artifacts.

As gerritcodereview image has reviewnotes plugin installed, update
the change-merged test in test_events_plugin.sh to account for an
extra event of type 'ref-updated' on 'refs/notes/review' which gets
fired when reviewnotes is installed. Also, fix the test script to
exit with the correct status.

Change-Id: I11c079c93dffb8703a4ae4e3cfb00e485385f58e