Cleanup the event capturing for diffs
Since there are several different types of captures happening in the
tests, qualify the methods which are related to the diffs. Also
explicitly cleanup the diff captures once the results have been
processed so that the resources are released before running any tests
which may be added after the diff (such as filtered event tests).
Finally, make it a bit clearer to which tests the diffs are intending to
applied to by starting the captures inside the specific test GROUP which
will be diffed.

Change-Id: I5bc8a95e24a84ab23c7f46a2824bea5f9a1d0121