Avoid streaming check on dropped events
Since dropped events don't need to be sent out to streamListeners, avoid
the extra overhead of checking to see if they need to be fired to the
streamListeners. This helps reduce the load on the event system as each
update check at a minimum requires reading the on disk "head" value.

Release-Notes: Dropped events now result in less IO
Change-Id: Ic87a53924dbdf4ce1b26450a5fb5fb46308c9a20